Added javadoc everywhere
Added java documentation in pretty much every single file, set up
Eclipse once more to be ready to work on the GUI again, since that’s
the next goal on the horizon.  TODO: Make the file path storage dynamic
and declared at launch, so it can be adjusted via the program in a
future release.