Fixed up the GUI and migrated the program to it
Rigged up the GUI almost completely.  Still missing dynamic file paths,
probably going to store that information in the app data folder of the
computer it is run on and have a file bar at the top to change storage
locations.  Also debating over interfacing with Powerlink in some way,
but it seems difficult to read the text of the makes of vehicles from
it.  Not entirely sure how to, honestly.