Restructured image loading to clean up code.
Trying to work on optimizing the code structure right now as it has to
be extremely dynamic to work with flexible file paths and work in more
environments than my own.  I’m planning to implement a user account
system of sorts on the SQL database that just identifies unique
computers and saves their configuration there.  All just conceptual at
the moment, but one day it will be a thing