Did some reformatting and centralized mouse adapter loading
I essentially moved all of the mouse listeners into a function and have
them all just passing the label and picture number to it.  This is more
unification for the planned dynamic file path system.  Next on the
chopping block is adding more image displays, perhaps a scrollable
gallery of sorts.