Login is now functional.  Committing this now in case I break everything.
Hey, the login works now that’s neat.  It is, however, a very insecure
method of logging in.  Don’t try this at home kids.  It’s being used on
a closed network with a bunch of people who can barely get around a
computer, so there isn’t much worry about the integrity being
compromised.