Formatting fixes, pricing frame centralized, implementation for quick sort
The purpose behind quick sort is rapidly sorting cars based on stock
number to folders based on the day they were indexed and in individual
folders based on stock number for easy migration to the website, where
they’ll be displayed by the week.  Perhaps it will support an archiving
system for older car pictures that are no longer being displayed on the
website, but as of right now that’s just speculation.