Began implementing the search function somewhat
Added a new framework for a list GUI that will display all results from
a search done.  I also started working on making the search reliably
find what year the car is in.  That should make it a lot easier to
narrow down where the car they’re searching for is in the tables.