Migrated some GUIs to be dialogs rather than JFrames. Other various modifications
The core change made was moving some windows be dialogs, so they stay
on top and function appropriately.  I also made the search engine
refine down to make if possible now, but there seems to be some bugs
I’ve got to work out with it still before it’s ship shape.